Refactor security details to multiple components, add security chart
