+++ package.json
@@ -5,6 +5,7 @@
   "homepage": ".",
   "dependencies": {
     "@patternfly/patternfly": "^4.185.1",
+    "@patternfly/react-charts": "^6.74.3",
     "@patternfly/react-core": "^4.202.16",
     "@patternfly/react-icons": "^4.53.16",
     "@patternfly/react-table": "^4.71.16",
