+++ src/resources/TagResource.ts
@@ -129,6 +129,7 @@ export async function getSecurityDetails(
     const response: AxiosResponse<SecurityDetailsResponse> = await axios.get(
       `/api/v1/repository/${org}/${repo}/manifest/${digest}/security?vulnerabilities=true`,
     );
+    console.log(response);
     return response.data;
   } catch (error: any) {
     throw new Error(`API error getting security details ${error.message}`);
