+++ src/tests/fake-db/ApiMock.ts
@@ -2,5 +2,6 @@ import {mock} from './MockAxios';
 
 import './data/login/login';
 import './data/user/user';
+import './data/repository/security';
 
 mock.onAny().passThrough();
