+++ webpack.dev.js
@@ -7,7 +7,7 @@ const PORT = process.env.PORT || "9000";
 
 module.exports = merge(common('development'), {
     mode: "development",
-    devtool: "eval-source-map",
+    devtool: "inline-source-map",
     devServer: {
         host: HOST,
         port: PORT,
